Rewrite video page to support blur at build time
