--- src/components/Video/VideoContent/VideoContent.module.scss
@@ -1,48 +1,0 @@
-.frame {
-  position: relative;
-  overflow: hidden;
-  padding: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.video {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  top: 0;
-  left: 0;
-  object-fit: cover;
-  opacity: 0;
-  transition: opacity 300ms ease-in;
-
-  .isLoaded & {
-    opacity: 1;
-  }
-}
-
-.brokenVideo {
-  width: 100%;
-  aspect-ratio: 16/9;
-  height: 100%;
-  background: var(--light-grey);
-}
-
-.placeholder {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  opacity: 0;
-  filter: blur(5px);
-  transform: scale(1.1);
-  position: absolute;
-  top: 0;
-  left: 0;
-  pointer-events: none;
-  transition: opacity 300ms ease-in;
-
-  .hasSizeData & {
-    opacity: 1;
-  }
-}
