+++ src/pages/video/[id]/index.tsx
@@ -1,14 +1,53 @@
 import { useRouter } from 'next/router';
 import Link from 'next/link';
-import { useCallback, useRef, useState } from 'react';
+import { useCallback, useMemo, useRef, useState } from 'react';
+import tinycolor from 'tinycolor2';
+import useSwr from 'swr';
+import classNames from 'classnames';
 import Video from '../../../components/Video/Video';
 import styles from './index.module.css';
 import Controls from '../../../components/Video/Controls/Controls';
+import { VideoData } from '../../../types/types';
+import {
+  getVideoDetailsById,
+  sanitiseVideoId,
+} from '../../../utils/videoUtils';
+import { getVideosList } from '../../../../netlify/functions/getVideosList';
+import { getVideoDetailsByIdOnServer } from '../../../server/methods';
 
-const VideoFocusModePage = () => {
+// Fetcher function that fetches data from getVideoData netlify function
+const fetcher = async (videoRef: string, video: VideoData) => {
+  if (video) {
+    return video;
+  }
+
+  // Some old videos are a full URL, rather than an ID
+  const id = sanitiseVideoId(videoRef);
+
+  const details = await getVideoDetailsById(id);
+
+  if (!details) {
+    console.error('Unable to retrieve video details for ID', id);
+
+    throw new Error('No details found for id');
+  }
+
+  return details;
+};
+
+type VideoFocusModePageProps = {
+  video: VideoData;
+};
+
+const VideoFocusModePage = ({ video }: VideoFocusModePageProps) => {
   const videoRef = useRef<HTMLVideoElement>();
   const router = useRouter();
+
   const { id, projectId } = router.query;
+
+  const { data, error } = useSwr(id, () => fetcher(id as string, video), {
+    fallbackData: video,
+  });
 
   const [isReady, setIsReady] = useState(false);
 
@@ -24,8 +63,29 @@ const VideoFocusModePage = () => {
     setIsReady(true);
   }, []);
 
+  const hasInvertedColors = useMemo(() => {
+    if (!data) {
+      return false;
+    }
+
+    const bgColor = tinycolor(data.blur.dominantColor);
+    const textColor = tinycolor('white');
+
+    return !tinycolor.isReadable(bgColor, textColor);
+  }, [data]);
+
+  const gridClassNames = classNames(styles.grid, {
+    [styles.invertedColors]: hasInvertedColors,
+  });
+
   return (
-    <div className={styles.grid}>
+    <div
+      className={gridClassNames}
+      style={{
+        backgroundColor: data.blur.dominantColor,
+        backgroundImage: `url(data:image/jpeg;base64,${data.blur.background})`,
+      }}
+    >
       <div className={styles.close}>
         {projectId ? (
           <Link href={`/project/${projectId}`}>{'View case study'}</Link>
@@ -35,15 +95,20 @@ const VideoFocusModePage = () => {
       </div>
 
       <div className={styles.video}>
-        <Video
-          className={styles.frame}
-          hasControls={false}
-          id={id as unknown as string}
-          isAutoplaying
-          onClick={handleClick}
-          onReady={handleReady}
-          ref={videoRef}
-        />
+        {error || !data ? (
+          <div className={`${styles.frame} ${styles.brokenVideo}`} />
+        ) : (
+          <Video
+            className={styles.frame}
+            hasControls={false}
+            isAutoplaying
+            isLooping
+            onClick={handleClick}
+            onReady={handleReady}
+            ref={videoRef}
+            video={data}
+          />
+        )}
       </div>
 
       {isReady && (
@@ -57,5 +122,28 @@ const VideoFocusModePage = () => {
     </div>
   );
 };
+
+export const getStaticProps = async ({ params }) => {
+  const video = await getVideoDetailsByIdOnServer(params.id);
+
+  return {
+    props: {
+      video,
+    },
+  };
+};
+
+export async function getStaticPaths() {
+  const videos = await getVideosList();
+
+  return {
+    fallback: false,
+    paths: videos.map(({ guid }) => ({
+      params: {
+        id: guid,
+      },
+    })),
+  };
+}
 
 export default VideoFocusModePage;
