+++ src/server/methods.ts
@@ -1,9 +1,13 @@
 /* Server side only imports */
 import { createBlurredImage } from '../../netlify/functions/getVideoThumbnail';
+import { VideoData } from '../types/types';
 import { getVideoDetailsById } from '../utils/videoUtils';
 
 export const getVideoDetailsByIdOnServer = async (id: string) => {
-  const videoDetails = await getVideoDetailsById(id);
+  const videoDetails = (await getVideoDetailsById(id)) as Omit<
+    VideoData,
+    'blur'
+  >;
 
   if (!videoDetails) {
     return null;
